Qt Designer and Pie Chart Test
Experimented with PyQt6 and Qt Designer, as well as learned the basics on how to make a table and a pie chart. These will all be important tools to build our UI. The code in pieChartTest.py is from the Qt documentation: https://doc.qt.io/qtforpython-6/examples/example_charts_piechart.html.