Built executables, updated README, created log file
Ready to submit for the Build and Test assignment. We now have an executable that can be found in the dist folder, as well as a failed build that can be found in the dist (failed) folder. I created a log file that shows the results of both these builds in the console. Also updated the readme to include more information in less space.