The user can now change a mod's priority level
Each row in the table now displays their mod's priority level (or "Ready") in a button that when clicked, shows a dropdown menu that allows the user to change that mod's priority level. Right now, the pie chart does not update, but the mod information and the color of the button does.

Additionally, I also added more data so we can ensure the UI works when scrolling.