UI tests and refactoring
DetailsWindow no longer creates the pyqt app and window, or have the example data in it. That has been moved to main.py. Additionally, all double underscore variables have been changed to single underscore variables.

I also created several tests for the UI. I plan to make more in the future, and use them to find the bug Andrew was struggling with.