More UI tests and remove mod button
Added more UI tests to tests.py, particularly focusing on actions the user can do (add a mod specifically through clicking the addModBtn, etc). Unfortunately, there's a strange bug with the last test that only pops up if the other UI tests are disabled. It's really strange, but everything working at the moment outside of this one test, so I'm going to push my progress now and fix it later.

Also, I added a button which lets you remove mods from the table. The button sits in a new 4th column of the table.