Cleaned up tests.py, minor refactoring
I never quite solved the issues with the dropdown test, but I got it to a point where it passes (as long as you don't change the modIndex), and has a lot of documentation about it. I also made some general changes to tests.py, including making the initDetailsView function global to that file. This removes some repeat code, since both of the UI test classes use it. Finally, I also made some minor changes across widgets.py and detailsWindow.py, including boundary checks in getter functions and replacing the centralWidget object with the mainWindow.