Adding pie chart tests and refactoring UI tests
Big news: all the tests work now! Even the dropdown one! It only took an additional 10 hours... (>18 total).

Even then, it's far from perfect. The tests need to be executed in a very specific order, and I have no idea why (beyond it being a PyQt6 or unittest problem). Fortunately, these issues do not effect the main program, and the tests pass in isolation. It's only an issue when they're run one after another. In the future, we may have to investigate other solutions, but for now, this works.

I've done some refactoring along the way, including moving all the tests to different files and adding setUp and tearDown functions. I plan to make some additional changes soon now that the system is working.