Added refresh button and selected version text input
In the top right corner of the screen, a new button has been added to refresh the table and pie chart. In the future, it will also refresh the API. Additionally, a new text field has been added next to it which allows the user to enter a selected version.