addModTextField uses placeholder text & tests now enter URL correctly
There were some issues after Andrew added input validation for API calls that caused some of the tests to fail. This was due to the tests inputting the URL without clearing the addModTextField first. This has been addressed. Additionally, the code to simulate adding a mod has been moved to a single function in DetailsWindow, rather than repeating it in each test.

In this process, I also changed the default text of the addModTextField to placeholder text, so the user doesn't have to delete it every time.