Back button and started add profile button
In the details view, there is now a back button that allows the user to return to the profile select view. To accomplish this, the QStackedWidget was turned into WindowManager class (which inherits from QStackedWidget).

I also started on the add profile button, but I disabled it for now because it doesn't work. I'll investigate more this weekend.