User can add profiles, refactoring
The user can now add mod profiles! When clicking the + button, they'll be given a dialog box to enter the name for the new profile, and then the layout will adjust accordingly. To accomplish this, I ripped out the old code for adding profile widgets and replaced it with a simpler QGridLayout.

With this, the UI for Mod Tracker's MVP is done!!!