Big performance increase, loading window, and refactoring
API calls now always happen on their own thread, drastically improving load times. The application launches faster and refreshes the profile faster. To prevent this from randomizing the order mods are displayed in, each mod now has a tablePosition attribute which is saved to the json file. This is used to sort the mod list. I also added a small loading window that opens when the app is launching to let the user know the app is doing something. The window has a gif, but unfortunately, the gif doesn't play until the app is done loading, at which point the window is closed anyway.

I changed the file structure to be less cluttered. Deprecated and temporary files made for testing and development have been moved to their own folder, and the test files have been moved to their own folder. Additionally, several functions have seen small refactoring. WindowManager now handles running load.createProfileList() and adding the results to ProfileSelectWindow, instead of main.py.