Curseforge support!
Curseforge is now properly supported by Mod Tracker! The process is slightly more involved, requiring multiple steps to sort the version list, as well as ensure what we are fetching is a mod, not a modpack or a world.

Now that Curseforge is properly supported, the refresh system will now use the service that the link originally came from. If you added the mod using a Mondrinth link, the data from Mondrinth will be used. Same with Curseforge.