App Icon and Profile Select Window Tests
Mod Tracker now has an app icon! It's a monkey wrench in Minecraft's style created by me. There were a few tweaks that needed to be made so the app icon displays on the taskbar and in the actual executable.

I also made 5 new tests for the Profile Select Window class. I will probably add a few more tests so all UI elements have proper tests, but this is acceptable for now.