mods.json backwards compatibility, inno setup script, updated README
If mods.json does not have the new tablePosition attirbute, the mod's table position will be sent to -1. If the tablePosition is negative, the mod list will be sorted alphabetically.

Added a script for making an installer for the executable with inno setup, as well as instructions for building the installer in the README.

Made some additional changes to the README, including to the installation instructions. Users will now be pointed to the releases page on GitHub, rather than the actions page. The screenshot should now display correctly, as it's getting it from a file, not a link. Made some other minor wording changes.