Fixed issue with tasker icon and the installer
The app's taskbar icon will now be correctly displayed as the app icon (instead of the default executable icon) on the taskbar when the executable is installed using the installer. Unfortunately, it still doesn't display correctly in the top left of the window. I'll hopefully find a fix for this in a future version of Mod Tracker.

Minor changes were also made to the README.

This is the final commit before the release of b2.0. The beta-2.0 branch will now be merged with main, and development will move to another branch.