Test text only shows when not running from main.py
In the mod table, the mod name text is handled by a label with invisible text under it for testing purposes. This text would sometimes still display on some systems, particularly in light mode. This behavior has been changed so that it only sets the text if the program was not started from main.py.