File reorganization
Files that are responsible for the desktop app are in the DesktopApp folder, similar to how files responsible for the web app are in the WebApp folder. Python files associated with backend tasks are now under the Backend folder. Images and assets have also been moved to the new assets folder. load.py has been renamed to loadFromJson.py, readJarFile.py to loadFromJar.py, server.py to pyServer.py, tests.py to runTests.py, and run.ps1 to runServers.ps1. Updated README and build-test.yml to reflect these changes.